backend/requirements.txt
# using python=3.11
notebook==7.0.5
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
zennit-crp==0.6.0